+++ src/DuckDB.EFCoreProvider/Properties/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 ﻿using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("DuckDB.EFCoreProvider.FunctionalTests")]
+[assembly: InternalsVisibleTo("DuckDB.EFCoreProvider.NTS")]