+++ src/DuckDB.EFCoreProvider/Storage/Internal/DuckDBTimestampTypeMapping.cs
@@ -111,7 +111,7 @@ public class DuckDBTimestampTypeMapping : RelationalTypeMapping
     /// <inheritdoc />
     protected override void ConfigureParameter(DbParameter parameter)
     {
-        ((DuckDBParameter)parameter).RemoveDollarSign();
+        ((DuckDBParameter)parameter).ConfigureNameAndMetadata(this);
         base.ConfigureParameter(parameter);
     }